🏷 Renamed MetricCallback to MetricsConfigurationCallback
