+++ anomalib/utils/callbacks/metrics_configuration.py
