+++ tests/pre_merge/utils/callbacks/visualizer_callback/dummy_lightning_model.py
@@ -10,7 +10,7 @@ from torch import nn
 from torch.utils.data import DataLoader, Dataset
 
 from anomalib.models.components import AnomalyModule
-from anomalib.utils.callbacks.metrics import MetricsConfigurationCallback
+from anomalib.utils.callbacks.metrics_configuration import MetricsConfigurationCallback
 from anomalib.utils.callbacks.visualizer_callback import VisualizerCallback
 from anomalib.utils.metrics import get_metrics
